Change maxCount to dropdown option
Enable to access all search results from header and footer dropdown functions.
