+++ test/integration/dropdown-option-spec.js
@@ -1,5 +1,6 @@
 import Textcomplete from '../../src/textcomplete';
 import Textarea from '../../src/textarea';
+import {CLASS_NAME} from '../../src/dropdown-item';
 
 const assert = require('power-assert');
 
@@ -12,10 +13,16 @@ describe('Dropdown options integration test', function () {
     textarea = new Textarea(textareaEl);
   });
 
-  function setup(option) {
+  function setup(option, strategy) {
     var textcomplete = new Textcomplete(textarea, { dropdown: option });
-    textcomplete.register([
-      {
+    textcomplete.register([strategy]);
+    return textcomplete;
+  }
+
+  describe('className', function () {
+    it('should be set as class attribute of dropdown element', function () {
+      var className = 'hello-world';
+      setup({ className }, {
         usernames: ['alice'],
         match: /(\w+)$/,
         search: function (term, callback) {
@@ -26,15 +33,7 @@ describe('Dropdown options integration test', function () {
         replace: function (username) {
           return `$1@${username} `;
         },
-      },
-    ]);
-    return textcomplete;
-  }
-
-  describe('className', function () {
-    it('should be set as class attribute of dropdown element', function () {
-      var className = 'hello-world';
-      textarea = setup({ className: className });
+      });
       textareaEl.value = '@a';
       var keyupEvent = document.createEvent('UIEvents');
       keyupEvent.initEvent('keyup', true, true);
@@ -48,7 +47,18 @@ describe('Dropdown options integration test', function () {
 
   describe('style', function () {
     it('should be set as style attribute of dropdown element', function () {
-      textarea = setup({ style: { backgroundColor: '#f0f' } });
+      setup({ style: { backgroundColor: '#f0f' } }, {
+        usernames: ['alice'],
+        match: /(\w+)$/,
+        search: function (term, callback) {
+          callback(this.usernames.filter((username) => {
+            return username.startsWith(term);
+          }));
+        },
+        replace: function (username) {
+          return `$1@${username} `;
+        },
+      });
       textareaEl.value = '@a';
       var keyupEvent = document.createEvent('UIEvents');
       keyupEvent.initEvent('keyup', true, true);
@@ -57,6 +67,27 @@ describe('Dropdown options integration test', function () {
 
       var dropdownEl = document.getElementsByClassName('textcomplete-dropdown')[0];
       assert.equal(dropdownEl.style.backgroundColor, 'rgb(255, 0, 255)');
+    });
+  });
+
+  describe('maxCount', function () {
+    it('should truncate the items of dropdown', function () {
+      var maxCount = 3;
+      setup({ maxCount }, {
+        match: /(\s|^)@(\w+)$/,
+        search: function (term, callback) { callback([0, 1, 2, 3, 4, 5, 6, 7, 8, 9]); },
+        replace: function (value) { return value; },
+      });
+
+      textareaEl.value = '@a';
+      textareaEl.selectionStart = textareaEl.selectionEnd = 2;
+      var keyupEvent = document.createEvent('UIEvents');
+      keyupEvent.initEvent('keyup', true, true);
+      keyupEvent.keyCode = 50;
+      textareaEl.dispatchEvent(keyupEvent);
+
+      var items = document.getElementsByClassName(CLASS_NAME);
+      assert.equal(items.length, maxCount);
     });
   });
 });
